refac(phoenix-ui): move url helpers to 'ContentHelpers', move `:index` of podcasts to the more approriate `:show` of networks, prefix partials with `-`
