+++ lib/radiator_web/templates/admin/network/index.html.eex
@@ -7,7 +7,7 @@
     <div class="flex justify-center md:justify-start flex-wrap">
       <%= for network <- @networks do %>
       <div class="w-48 flex flex-col mr-8 mb-10">
-        <%= link to: Routes.admin_network_podcast_path(@conn, :index, network.id),
+        <%= link to: Routes.admin_network_path(@conn, :show, network.id),
         class: "no-underline text-black w-48 bg-white rounded flex flex-col shadow mr-8 mb-8" do %>
           
           <div
